:sparkles: Adds assume role on entrypoint
BLAC-1112

During the implementation of our serverless infrastructure we stumbled upon an
issue with drone-s3. On this plugin we're not being able to assume the current
account role, so it is always deploying as root.

On this commit I've changed the entrypoint to assume it if no other role was provided.

Now, we should be able to deploy as the right role.